Added batch put test. In table extension added option to delete attributes with empty values.

+++ Stratosphere/Stratosphere/Table/TableExtension.cs
@@ -18,15 +18,17 @@ namespace Stratosphere.Table
         }
 
         public static void Set<T>(this ITable table, string itemName, T itemData)
-            where T : IEnumerable<KeyValuePair<string, string>>, new()
+            where T : IEnumerable<KeyValuePair<string, string>>
+        {
+            Set(table, itemName, itemData, true);
+        }
+
+        public static void Set<T>(this ITable table, string itemName, T itemData, bool deleteEmpty)
+            where T : IEnumerable<KeyValuePair<string, string>>
         {
             if (!string.IsNullOrEmpty(itemName))
             {
-                var putAttributes = itemData.Where(
-                    pair => !string.IsNullOrEmpty(pair.Value)).ToArray();
-
-                string[] deleteAttributeNames = itemData.Where(
-                    pair => string.IsNullOrEmpty(pair.Value)).Select(pair => pair.Key).ToArray();
+                var putAttributes = itemData.Where(pair => !string.IsNullOrEmpty(pair.Key)).ToArray();
 
                 if (putAttributes.Length != 0)
                 {
@@ -34,34 +36,42 @@ namespace Stratosphere.Table
                     {
                         foreach (var attribute in putAttributes)
                         {
-                            w.ReplaceAttribute(attribute.Key, attribute.Value);
+                            w.ReplaceAttribute(attribute.Key, attribute.Value != null ? attribute.Value : string.Empty);
                         }
                     });
                 }
 
-                if (deleteAttributeNames.Length != 0)
+                if (deleteEmpty)
                 {
-                    table.Delete(itemName, w =>
+                    string[] deleteAttributeNames = itemData.Where(
+                        pair => !string.IsNullOrEmpty(pair.Key) && string.IsNullOrEmpty(pair.Value)).Select(pair => pair.Key).ToArray();
+
+                    if (deleteAttributeNames.Length != 0)
                     {
-                        foreach (string name in deleteAttributeNames)
+                        table.Delete(itemName, w =>
                         {
-                            w.DeleteAttribute(name);
-                        }
-                    });
+                            foreach (string name in deleteAttributeNames)
+                            {
+                                w.DeleteAttribute(name);
+                            }
+                        });
+                    }
                 }
             }
         }
 
         public static void Set<T>(this ITable table, IEnumerable<KeyValuePair<string, T>> items)
-            where T : IEnumerable<KeyValuePair<string, string>>, new()
+            where T : IEnumerable<KeyValuePair<string, string>>
         {
-            var batchPutItems = items.Select(item =>
-                new KeyValuePair<string, KeyValuePair<string, string>[]>(item.Key, item.Value.Where(
-                    pair => !string.IsNullOrEmpty(pair.Value)).ToArray())).Where(item => item.Value.Length != 0).ToArray();
+            Set(table, items, true);
+        }
 
-            var deleteItems = items.Select(item =>
-                 new KeyValuePair<string, string[]>(item.Key, item.Value.Where(
-                     pair => string.IsNullOrEmpty(pair.Value)).Select(pair => pair.Key).ToArray())).Where(item => item.Value.Length != 0).ToArray();
+        public static void Set<T>(this ITable table, IEnumerable<KeyValuePair<string, T>> items, bool deleteEmpty)
+            where T : IEnumerable<KeyValuePair<string, string>>
+        {
+            var batchPutItems = items.Where(item => !string.IsNullOrEmpty(item.Key)).Select(item =>
+                new KeyValuePair<string, KeyValuePair<string, string>[]>(item.Key, item.Value.Where(pair => 
+                    !string.IsNullOrEmpty(pair.Key)).ToArray())).Where(item => item.Value.Length != 0).ToArray();
 
             if (batchPutItems.Length != 0)
             {
@@ -71,25 +81,33 @@ namespace Stratosphere.Table
                     {
                         foreach (var attribute in item.Value)
                         {
-                            w.ReplaceAttribute(item.Key, attribute.Key, attribute.Value);
+                            w.ReplaceAttribute(item.Key, attribute.Key, attribute.Value != null ? attribute.Value : string.Empty);
                         }
                     }
                 });
             }
 
-            if (deleteItems.Length != 0)
+            if (deleteEmpty)
             {
-                foreach (var item in deleteItems)
+                var deleteItems = items.Where(item => !string.IsNullOrEmpty(item.Key)).Select(item =>
+                     new KeyValuePair<string, string[]>(item.Key, item.Value.Where(pair => 
+                         !string.IsNullOrEmpty(pair.Key) && string.IsNullOrEmpty(pair.Value)).Select(pair => pair.Key).ToArray())).Where(item => 
+                             item.Value.Length != 0).ToArray();
+
+                if (deleteItems.Length != 0)
                 {
-                    if (item.Value.Length != 0)
+                    foreach (var item in deleteItems)
                     {
-                        table.Delete(item.Key, w =>
+                        if (item.Value.Length != 0)
                         {
-                            foreach (var name in item.Value)
+                            table.Delete(item.Key, w =>
                             {
-                                w.DeleteAttribute(name);
-                            }
-                        });
+                                foreach (var name in item.Value)
+                                {
+                                    w.DeleteAttribute(name);
+                                }
+                            });
+                        }
                     }
                 }
             }
@@ -147,7 +165,10 @@ namespace Stratosphere.Table
                         }
                     }
 
-                    itemData[reader.AttributeName] = reader.AttributeValue;
+                    if (!string.IsNullOrEmpty(reader.AttributeValue))
+                    {
+                        itemData[reader.AttributeName] = reader.AttributeValue;
+                    }
                 }
 
                 if (itemName != null)
